# src/v2/docs/openapi.yaml
openapi: 3.0.0
info:
  title: TMJApp API V2
  version: 2.0.0
  description: API v2 organizada por visoes (admin, passenger, driver).
servers:
  - url: http://localhost:3000/api/v2
  - url: http://api.tmjapp.com.br/api/v2
...
paths:
...
  /admin/products/{id}:
    get:
      tags: [Admin V2]
      summary: Operações em produto específico
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operação realizada com sucesso
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Erro de validação ou regra de negócio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    put:
      tags: [Admin V2]
      summary: Operações em produto específico
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Operação realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
        '400':
          description: Erro de validação ou regra de negócio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      tags: [Admin V2]
      summary: Operações em produto específico
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operação realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
        '400':
          description: Erro de validação ou regra de negócio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /admin/tarifas:
    get:
      tags: [Admin V2]
      summary: Gerenciar tarifas globais
      responses:
        '200':
          description: Operação realizada com sucesso
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Erro de validação ou regra de negócio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    post:
      tags: [Admin V2]
      summary: Gerenciar tarifas globais
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Operação realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
        '400':
          description: Erro de validação ou regra de negócio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /admin/tarifas/{id}:
    put:
      tags: [Admin V2]
      summary: Operações em tarifa específica
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Operação realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
        '400':
          description: Erro de validação ou regra de negócio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      tags: [Admin V2]
      summary: Operações em tarifa específica
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operação realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
        '400':
          description: Erro de validação ou regra de negócio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    SuccessMessageResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    ApiErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        code:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true
        traceId:
          type: string
security:
  - bearerAuth: []